Rename error.rs -> results.rs
