+++ src/gg20/mnemonic/results.rs
